Update AssemblyVersion and modify App.xaml handling
Updated AssemblyVersion from 1.1.3 to 1.1.4 to reflect a new version. Adjusted the handling of `App.xaml` by removing its `ApplicationDefinition` and including it as a `<Page>` element in the project file. This change likely impacts how the file is compiled. Other project settings, such as target framework, output type, and WPF usage, remain unchanged.